Sync bytes setup when trying to read id
READ ID = connecting to the HIL device. If the HIL device is on an active CAN bus, it may be trying to send a lot of CAN message data async over the Serial, making the call-response system of READ_ID hard to execute. Instead 4 SYNC BYTES are send, and READ ID tries to search multiple serial chunks for the sync bytes + id response. Doing this could cause the Python to read part of a CAN message from the HIL device, throwing off the entire alignment of the parser. As such, any bytes after the sync bytes + id are saved and passed to the threaded serial.